+++ Source/SwiftLintBuiltInRules/Rules/RuleConfigurations/TestCaseAccessibilityConfiguration.swift
@@ -1,8 +1,5 @@
 import SwiftLintCore
 
-// swiftlint:disable:next blanket_disable_command
-// swiftlint:disable let_var_whitespace
-
 @AutoApply
 struct TestCaseAccessibilityConfiguration: SeverityBasedRuleConfiguration {
     typealias Parent = TestCaseAccessibilityRule