+++ Source/swiftlint/Helpers/LintOrAnalyzeArguments.swift
@@ -25,7 +25,7 @@ struct LintOrAnalyzeArguments: ParsableArguments {
         Should reformat the Swift files using the same mechanism used by Xcode (via SourceKit).
         Only applied with `--fix`/`--autocorrect`.
         """)
-    var format = false // swiftlint:disable:this let_var_whitespace - the multiline @Flag breaks us
+    var format = false
     @Flag(help: "Use an alternative algorithm to exclude paths for `excluded`, which may be faster in some cases.")
     var useAlternativeExcluding = false
     @Flag(help: "Read SCRIPT_INPUT_FILE* environment variables as files.")